Skip unavailable tests on Python 3.8 and ensure expected failures fail

+++ recipe/run_test.sh
@@ -1,6 +1,5 @@
 #!/usr/bin/env bash
-
-set -x
+set -ex
 
 echo "CONDA_PY:$CONDA_PY"
 export CONDA_PY=`python -c "import sys;print('%s%s'%sys.version_info[:2])"`
@@ -10,11 +9,17 @@ gdb -batch -ex "run" -ex "py-bt" --args python "$RECIPE_DIR/testing/process_to_d
 if [[ "$CONDA_PY" != "27" ]]; then
     grep "built-in method kill" gdb_output
 fi
-# Unfortunately python 3.6 package does not have enough debug info for py-bt
-if [[ "$CONDA_PY" != "36" && "$CONDA_PY" != "27" ]]; then
+# Unfortunately some python packages do not have enough debug info for py-bt
+if [[ "$CONDA_PY" != "36" && "$CONDA_PY" != "27" && "$CONDA_PY" != "38" ]]; then
     grep "line 3" gdb_output
     grep "process_to_debug.py" gdb_output
     grep 'os.kill(os.getpid(), signal.SIGSEGV)' gdb_output
+else
+    if grep "line 3" gdb_output; then
+        echo "This test was expected to fail due to missing debug info in python"
+        echo "As it passed the test should be re-enabled"
+        exit 1
+    fi
 fi
 
 grep "Program received signal SIGSEGV" gdb_output
